+++ .gitignore
@@ -26,3 +26,4 @@ go.work.sum
 
 .vscode/
 .vscode/*
+coverage.txt
